Split logic for finding ground into separate class
Made a dedicated class to house the logic of finding the 'ground'
position that a corpse will attempt to spawn at for a given set of 2D
coordinates.

Made the constants for min/max height of the world, and search radius more configurable.

Reduced the dimensions of the unit test world to speed up unit testing.