Update corpse position logic to account for spikes
Added new configuration parameter to set whether or not corpses should
be allowed to spawn on top of spikes.

Altered corpse positioning logic to use 2 delegates, one to check if the
ground block is stable, and another to check if the block above it, is a
valid spawn point.

Updated unit tests.